Update website metadata and backend API endpoints
Updated both metadata in index.html and API endpoints in App.js file. The reason for this is to enhance the nature of the website relevancy and as well its interaction with the API endpoints. Modified frontend/public/index.html to change page description and title to suit the new application's requirements. Backend API endpoints in frontend/src/App.js incorporated now an environment variable for dynamic assignment. This better facilitates different environments for development, testing, and production.