Add password length check and improve error alert messages
Added a length check condition for the password to ensure the entered password is between 5 to 10 characters. This is to improve password validation. Adjusted error messaging in catch blocks to let user know specifically when the entered password is wrong. Header content was rearranged to improve layout readability, making it more user friendly.